requirements.txt
flask==0.9
redis
gevent-socketio

